Fixed logic to build menu - task #809
